(refactor): implement adapter pattern for retrieving the zaaktypen choices inside form settings
