+++ src/GravityForms/GravityFormsFormSettings.php
@@ -2,139 +2,15 @@
 
 namespace OWC\Zaaksysteem\GravityForms;
 
-use Exception;
-use OWC\Zaaksysteem\Contracts\Client;
-use OWC\Zaaksysteem\Endpoints\Filter\ResultaattypenFilter;
-use OWC\Zaaksysteem\Entities\Zaaktype;
+use OWC\Zaaksysteem\GravityForms\ZaaktypenFormSettings\Clients\DecosClient;
+use OWC\Zaaksysteem\GravityForms\ZaaktypenFormSettings\Clients\OpenZaakClient;
+use OWC\Zaaksysteem\GravityForms\ZaaktypenFormSettings\Clients\RxMissionClient;
+use OWC\Zaaksysteem\GravityForms\ZaaktypenFormSettings\Clients\XxllncClient;
 use OWC\Zaaksysteem\Resolvers\ContainerResolver;
-
-use OWC\Zaaksysteem\Support\Collection;
-
-use function Yard\ConfigExpander\Foundation\Helpers\value;
 
 class GravityFormsFormSettings
 {
     protected string $prefix = OWC_GZ_PLUGIN_SLUG;
-
-    /**
-     * Get a list of related 'zaaktypen' from Open Zaak.
-     */
-    public function getTypesOpenZaak(): array
-    {
-        try {
-            return $this->getTypesByClient(ContainerResolver::make()->getApiClient('openzaak'));
-        } catch(Exception $e) {
-            return $this->handleNoChoices();
-        }
-    }
-
-    /**
-     * Get a list of related 'zaaktypen' from Decos Join.
-     */
-    public function getTypesDecosJoin(): array
-    {
-        try {
-            return $this->getTypesByClient(ContainerResolver::make()->getApiClient('decos'));
-        } catch(Exception $e) {
-            return $this->handleNoChoices();
-        }
-    }
-
-    /**
-     * Get a list of related 'zaaktypen' from Rx.Mission.
-     */
-    public function getTypesRxMission(): array
-    {
-        try {
-            return $this->getTypesByClient(ContainerResolver::make()->getApiClient('rx-mission'));
-        } catch(Exception $e) {
-            return $this->handleNoChoices();
-        }
-    }
-
-    /**
-     * Get a list of related 'zaaktypen' from Rx.Mission.
-     */
-    public function getTypesXxllnc(): array
-    {
-        try {
-            return $this->getTypesByClient(ContainerResolver::make()->getApiClient('xxllnc'));
-        } catch(Exception $e) {
-            return $this->handleNoChoices();
-        }
-    }
-
-    /**
-     * Return types by client, includes pagination.
-     */
-    protected function getTypesByClient(Client $client): array
-    {
-        $transientKey = sprintf('%s-form-settings-zaaktypen', $client->getClientNamePretty());
-        $types = get_transient($transientKey);
-
-        if (is_array($types) && $types) {
-            return $types;
-        }
-
-        $zaaktypen = $this->getTypes($client);
-        $types = $this->prepareTypes($zaaktypen);
-
-        set_transient($transientKey, $types, 500);
-
-        return $types;
-    }
-
-    protected function getTypes(Client $client): array
-    {
-        $page = 1;
-        $zaaktypen = [];
-        $requestException = '';
-
-        while ($page) {
-            try {
-                $result = $client->zaaktypen()->all((new ResultaattypenFilter())->page($page));
-            } catch (Exception $e) {
-                $requestException = $e->getMessage();
-
-                break;
-            }
-
-            $zaaktypen = array_merge($zaaktypen, $result->all());
-            $page = $result->pageMeta()->getNextPageNumber();
-        }
-
-        if (empty($zaaktypen)) {
-            $exceptionMessage = 'No zaaktypen found.';
-
-            if (! empty($requestException)) {
-                $exceptionMessage = sprintf('%s %s', $exceptionMessage, $requestException);
-            }
-
-            throw new Exception($exceptionMessage);
-        }
-
-        return $zaaktypen;
-    }
-
-    protected function prepareTypes(array $zaaktypen): array
-    {
-        return (array) Collection::collect($zaaktypen)->map(function (Zaaktype $zaaktype) {
-            return [
-                'name' => $zaaktype->identificatie,
-                'label' => "{$zaaktype->omschrijving} ({$zaaktype->identificatie})",
-                'value' => $zaaktype->identificatie,
-            ];
-        })->all();
-    }
-
-    protected function handleNoChoices(): array
-    {
-        return [
-            [
-                'label' => __('Unable to retrieve "Zaak" types provided by selected supplier.', 'owc-gravityforms-zaaksysteem'),
-            ],
-        ];
-    }
 
     /**
      * Add form based settings.
@@ -178,7 +54,6 @@ class GravityFormsFormSettings
                         ],
                     ],
                 ],
-                // TODO: verify if there is a way to actively get the selected value without a save and without custom JS.
                 [
                     'name'    => "{$this->prefix}-form-setting-openzaak-identifier",
                     'type'    => 'select',
@@ -192,7 +67,7 @@ class GravityFormsFormSettings
                             ],
                         ],
                     ],
-                    'choices' => $this->getTypesOpenZaak(),
+                    'choices' => (new OpenZaakClient(ContainerResolver::make()->getApiClient('openzaak')))->zaaktypen(),
                 ],
                 [
                     'name'    => "{$this->prefix}-form-setting-decos-join-identifier",
@@ -207,7 +82,7 @@ class GravityFormsFormSettings
                             ],
                         ],
                     ],
-                    'choices' => $this->getTypesDecosJoin(),
+                    'choices' => (new DecosClient(ContainerResolver::make()->getApiClient('decos')))->zaaktypen(),
                 ],
                 [
                     'name'    => "{$this->prefix}-form-setting-rx-mission-identifier",
@@ -222,7 +97,7 @@ class GravityFormsFormSettings
                             ],
                         ],
                     ],
-                    'choices' => $this->getTypesRxMission(),
+                    'choices' => (new RxMissionClient(ContainerResolver::make()->getApiClient('rx-mission')))->zaaktypen(),
                 ],
                 [
                     'name'    => "{$this->prefix}-form-setting-xxllnc-identifier",
@@ -237,7 +112,7 @@ class GravityFormsFormSettings
                             ],
                         ],
                     ],
-                    'choices' => $this->getTypesXxllnc(),
+                    'choices' => (new XxllncClient(ContainerResolver::make()->getApiClient('xxllnc')))->zaaktypen(),
                 ],
             ],
         ];
