+++ src/GravityForms/ZaaktypenFormSettings/Clients/ClientInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace OWC\Zaaksysteem\GravityForms\ZaaktypenFormSettings\Clients;
+
+interface ClientInterface
+{
+    public function zaaktypen(): array;
+    public function getClientNamePretty(): string;
+}
