+++ src/GravityForms/ZaaktypenFormSettings/Clients/DecosClient.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace OWC\Zaaksysteem\GravityForms\ZaaktypenFormSettings\Clients;
+
+use OWC\Zaaksysteem\GravityForms\ZaaktypenFormSettings\Adapters\ClientAdapter;
+
+class DecosClient extends ClientAdapter
+{
+    //
+}
