+++ src/GravityForms/ZaaktypenFormSettings/Clients/XxllncClient.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace OWC\Zaaksysteem\GravityForms\ZaaktypenFormSettings\Clients;
+
+use OWC\Zaaksysteem\GravityForms\ZaaktypenFormSettings\Adapters\ClientAdapter;
+
+class XxllncClient extends ClientAdapter
+{
+    //
+}
